+++ data_engines/__init__.py
@@ -1,7 +1,7 @@
 # Bring pertinent classes in to the namespace of this package
-from coder_engine.coder_engine import CoderEngine
-from coder_engine.language_engine import LanguageEngine
-from coder_engine.mongo_engine import MongoEngine
+from data_engines.coder_engine import CoderEngine
+from data_engines.language_engine import LanguageEngine
+from data_engines.mongo_engine import MongoEngine
 
 # Explicitly name what files are in this package
 __all__ = ['mongo_engine', 'language_engine', 'coder_engine']
